tests/wpt/mozilla/tests/bluetooth/getPrimaryServices/disconnected-device-with-uuid.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}]
    })
    .then(device => promise_rejects_dom(t, 'NetworkError', device.gatt.getPrimaryServices(heart_rate.name)));
}, 'getPrimaryServices called before connecting. Reject with NetworkError.');
</script>
